backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
cryptography==41.0.4
supabase==2.0.3
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.0.1
email-validator==2.0.0
Flask-Mail==0.9.1
python-dateutil==2.8.2
